Moved SkinConfiguration + sort package imports
