--- api/src/main/java/net/md_5/bungee/api/config/SkinConfiguration.java
@@ -1,23 +1,0 @@
-package net.md_5.bungee.api.config;
-
-/**
- * Represents a player's skin settings.
- */
-public interface SkinConfiguration
-{
-
-    boolean hasCape();
-
-    boolean hasJacket();
-
-    boolean hasLeftSleeve();
-
-    boolean hasRightSleeve();
-
-    boolean hasLeftPants();
-
-    boolean hasRightPants();
-
-    boolean hasHat();
-
-}
